# public/openapi.yaml
openapi: 3.0.0
info:
  title: Gestión de Vehículos API
  description: API para la administración de flotas de vehículos, conductores, trayectos y mantenimientos.
  version: 1.1.0
servers:
  - url: /api
    description: Servidor Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Vehiculo:
      type: object
      properties:
        matricula:
          type: string
        marca:
          type: string
        modelo:
          type: string
        fechaCompra:
          type: string
          format: date-time
        anyosAntiguedad:
          type: integer
        tipo:
          type: string
        kilometrosTotales:
          type: number
        alimentacion:
          type: string
        precio:
          type: number
        nuevo:
          type: boolean
        gastoPorKm:
          type: number
        imagenes:
          type: array
          description: "Format depends on the method (POST uses object with id, PATCH uses URL string)"
          items:
            anyOf:
              - type: object
                properties:
                  id:
                    type: integer
              - type: string
                format: uri
        conductorDni:
          type: string
          nullable: true
        averiaId:
          type: integer
          nullable: true
      required:
        - matricula
        - marca
        - modelo
        - fechaCompra
        - anyosAntiguedad
        - tipo
        - kilometrosTotales
        - alimentacion
        - nuevo
        - gastoPorKm
...
paths:
...
  /vehiculos:
    get:
      summary: Listar vehículos
      tags:
        - Vehiculos
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de vehículos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehiculo'
    post:
      summary: Crear un vehículo
      tags:
        - Vehiculos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehiculo'
            example:
              matricula: "1234ABC"
              marca: "Toyota"
              modelo: "Corolla"
              fechaCompra: "2023-01-01T00:00:00Z"
              anyosAntiguedad: 1
              tipo: "Turismo"
              kilometrosTotales: 1000
              alimentacion: "Gasolina"
              precio: 25000
              nuevo: true
              gastoPorKm: 0.15
              imagenes:
                - id: 1
                - id: 2
      responses:
        '201':
          description: Vehículo creado exitosamente
        '401':
          description: No autorizado
        '409':
          description: Matrícula o marca ya existen
...
  /files:
    get:
      summary: Listar recursos de Cloudinary (Imágenes guardadas en BD)
      tags:
        - Files
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de imágenes
    post:
      summary: Guardar referencia de imagen en BD
      tags:
        - Files
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                nombre:
                  type: string
              required: [url, nombre]
      responses:
        '201':
          description: Referencia guardada
...